Copy Unify Output

Copy the output generated by the unify script, to the setup directory, and symlink a few things. Note that recount3 is expecting human or mouse, with the relevant annotation. If you are using this, you could have generated anything you wanted (I build this to handle zebrafish using MZ11). Therefore, it copies the files, then symlinks them to the names that recount3 expects.

rc3_copy_unify_outputs(
  rc3_directory = NULL,
  short_id = "test",
  unify_directory = NULL,
  references_directory = NULL,
  organism = "human",
  annotation = "G026"
)

Arguments

rc3_directory

where you want things copied to (sort of)

short_id

the short id of the project

unify_directory

where the unify outputs are stored

references_directory

where the reference data is stored

organism

what organism is the data (for real or pretend)

annotation

what annotations are we interested in (for real or pretend)

Value

output of the "tree" command on rc3_directory