run recount-pump and recount-unify

Given the parameters, attempts to create and run the commands to carry out the runs of recount-pump and recount-unify on a set of samples.

Usage

rc3_run_pump_unify(
  fasta = getwd(),
  outputs = fs::path_dir(fs::path_expand(fasta)),
  monorail = "monorail-external",
  recount_pump = "recount-pump_1.1.3.sif",
  recount_unify = "recount-unify_1.1.1.sif",
  reference_path = ".",
  reference = "hg38",
  studyid = "other1",
  shortid = "test",
  ncore = 1,
  run = "both",
  delete_oldpump = "yes",
  run_or_show = "run"
)

Arguments

fasta

the directory with the input samples to be run

outputs

where to keep outputs, assumed to be the parent of fasta

monorail

where is the monorail repository

recount_pump

where is the recount-pump sif image

recount_unify

where is the recount-unify sif image

studyid

the study id

shortid

short id

ncore

how many cores to use

run

which parts of the workflow to run? (both, pump, unify)

delete_oldpump

should the old recount-pump results be deleted and regenerated?

run_or_show

do you want the commands shown, or actually run?

Value

unify output directory