run recount-pump and recount-unify

Given the parameters, attempts to create and run the commands to carry out the runs of recount-pump and recount-unify on a set of samples.

rc3_run_pump_unify(
  fasta = getwd(),
  outputs = fs::path_dir(fs::path_expand(fasta)),
  monorail = "monorail-external",
  recount_pump = "recount-pump_1.1.3.sif",
  recount_unify = "recount-pump_1.1.3.sif",
  reference_path = ".",
  reference = "hg38",
  studyid = "other1",
  shortid = "test",
  ncore = 1
)

Arguments

fasta

the directory with the input samples to be run

outputs

where to keep outputs, assumed to be the parent of fasta

monorail

where is the monorail repository

recount_pump

where is the recount-pump sif image

recount_unify

where is the recount-unify sif image

studyid

the study id

shortid

short id

ncore

how many cores to use

Value

unify output directory